Add elliptic package, file renaming and minor fixes.
