+++ package.json
@@ -14,7 +14,7 @@
     "test": "npm run test:node && npm run test:browser",
     "test:browser": "npm run build:dist && karma start karma.conf.js",
     "test:node": "npm run build:dist && istanbul test mocha -- --reporter spec",
-    "build:dist": "babel index.js crypto.js lib/* --source-root ./ -d ./dist",
+    "build:dist": "babel index.js secp256k1-adapter.js secp256k1-lib/* --source-root ./ -d ./dist",
     "build:docs": "documentation build ./index.js --github --sort-order='alpha' -f md > ./docs/index.md"
   },
   "repository": {
@@ -106,6 +106,7 @@
     "bn.js": "^4.11.0",
     "create-hash": "^1.1.2",
     "ethereum-cryptography": "^0.1.3",
+    "elliptic": "^6.5.2",
     "ethjs-util": "^0.1.3",
     "rlp": "^2.0.0",
     "safe-buffer": "^5.1.1"
